Simplify algorithm - uses string rather than case classes
This is to avoid having the information stored in a case class. This
poses a problem when you need to sum the values of a list, you need
to take the head of the list to extract the information. Better to
have this as meta data and keep the type as a string (although this
part could have been kept as a case object for example).